pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "aigravscode"
version = "0.0.1"
authors = [
  { name="Example Author", email="author@example.com" },
]
description = "A modern, multi-stack project starter template"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]

[project.urls]
"Homepage" = "https://github.com/0nkar/aigravscode"
"Bug Tracker" = "https://github.com/0nkar/aigravscode/issues"

[project.scripts]
aigravscode = "aigravscode.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 88

[tool.ruff]
line-length = 88
target-version = "py38"
select = ["E", "F", "I"]
ignore = []
